+++ assets/components/ui/Button.tsx
@@ -1,9 +1,17 @@
 import type { FunctionComponent, ReactNode } from 'react';
 
-const Button: FunctionComponent<ButtonProps> = ({ outlined, children }) => {
+const Button: FunctionComponent<ButtonProps> = ({
+    outlined,
+    children,
+    block,
+    onClick,
+}) => {
     return (
         <button
-            className={`button button-component${outlined ? '-outlined' : ''}`}
+            className={`button button-component${outlined ? '-outlined' : ''} ${
+                block && 'button-component-block'
+            }`}
+            onClick={onClick}
         >
             {children}
         </button>
@@ -15,4 +23,6 @@ export default Button;
 interface ButtonProps {
     children: ReactNode;
     outlined?: boolean;
+    block?: boolean;
+    onClick?: () => void;
 }