+++ assets/components/ui/index.ts
@@ -1,5 +1,6 @@
 import SectionTitle from './SectionTitle';
 import FormInput from './FormInput';
 import Button from './Button';
+import ViewData from './ViewData';
 
-export { SectionTitle, FormInput, Button };
+export { SectionTitle, FormInput, Button, ViewData };