+++ utils/index.ts
@@ -1,0 +1,3 @@
+import makeArray from './makeArray';
+
+export { makeArray };